requirements.txt
pygithub
slackweb
configargparse
